requirements.txt
langchain 
chromadb 
tabulate 
sqlalchemy 
sqlalchemy-bigquery 
google-cloud-bigquery
google-cloud-aiplatform
Flask==2.0.3
gunicorn==20.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
pandas==1.3.1
textblob==0.17.1
scikit-learn
datetime
datalab
arrow
db_dtypes
pandas-gbq
flask_mail
flask_cors
google-api-core
